Rough in CRUD for Users Resource
At this point I can successfully create, list and delete
users.

- add driver for Postgres

Notes:
- When using .Scan() it was necessary to scan ALL table
fields; otherwise, there would be an error.
- In production, I would not put the db IP, UN and password
in code that is committed to Git.
- I used plain JS to work the DELETE just to keep things moving;
it would take a bit of time to set up a proper front end
build system with webpack or similar, and I'd rather tackle
that later.
- next up will be to clean up the code with better error
handling, etc. and properly encrypt the password.