WIP - Rough in Basic Auth - Login/Logout/Register/Dashboard
This is really really roughed in.
Still to do: password encryption, reset...

I added the auth.go file so I could just have a
blank slate to build this out. I'll decide later
on how I want to structure this all. I'll fight
the urge to over-think it.

As I built this out, I struggled with the following:
- naming routes and methods (login, login-user, etc... bleh)
- passing around db / log references
- variables used for scanning db records, creating
the authenticated user struct, etc. It just felt
like a lot to keep track off.
- deciding which funcs to make methods on structs
and which to leave as plain methods. eg: making
funcs methods on the authController struct made
it easy to share the db and log references,
but I wonder if this is the best way to do that.

I will be doing some clean up next to see if I can't
make this all easier to follow.