WIP - Add Features, Update Projects
- add features routes and pages, etc.
- update project sql as SELECT * will
cause errors once the db schema changes;
it's better to be explicit with SELECTs
to avoid those errors. I'll have to look
for other places I made this mistake.
- Add Features and Project to their related
structs. I'm not sure if this is idiomatic Go,
but I'm used to these nested relationships
with Laravel, and it makes sense to me.
- simplify the use of Scan() to write directly
to the struct fields.
- update updated_at fields on update
- add links to various pages to make navigation
a bit easier.

Nailing down the routes was tricky. httprouter
is much pickier with routes than Laravel, and I
found myself dealing with a route conflict error
more than a few times. Part of the problem was
just using :id for a param. This caused conflicts,
so I settled on feature_id, project_id, convention.
I'm fairly content with what I ended up with.
I'll have to look at the others to make sure I'm
staying consistent.

Laravel's Eloquent has spoiled me a bit;
I looked around for the best way to duplicate
the result I get with Laravel's query()->with()
clause. For my current case of just wanting a
single Project with related features my use
of 2 separate queris seems ok, but I'd like
to know how to do this with raw SQL, so
that will take a bit more research.