Features - Remove show() redirect to new / create()
I changed this route and forgot to update this.
This is a copy-paste error from the projects.go
file where projects.show handles :project_id
and redirects to the create form when :project_id
is "new" and not an actual project_id. I did this
b/c of the way httprouter works.