Add Stories Resource
- this is pretty much a copy / paste job
from doing the Features resource with
one addition - Assignee relationship.

Unlike projects and features, stories can
be assigned to a user, though not always.
Since the Assignee can be a user struct or
null, I had to use sql.NullString for the
assignee_name instead of User.Name (string)
and sql.NullInt64 for the ID.