Return ExecuteTemplate(error) from view.exec
I had first returned an error with make() but
I thought using Must() around ParseFiles and
returning the error from ExecuteTemplate
looked closer to how I see html templates used
in the wild. I'm still not sure if this is the
best way to approach this, but following
the apparent trend seems like a reasonable way to
start.