Add New Project button to projects templates
- add content_menu template

If content_menu isn't defined an error is thrown.
I couldn't find a way to check if the template
exists before calling {{template "content_menu"}},
so to avoid an error I defined the template in the
layout file. As long as the final content page is
parsed after the layout, which is always is,
then the last definition for the template will
be the one used.