Add Stories to Feature Page
This is much like what I did to show
features on a project page.
There is a lot of repeat code with the *PageModule.js,
but I'm not too worried about it at this point;
there isn't a lot of code here, and I'm about
to add a list of bugs to the feature pages, so keeping
these modules separate will probably make things
easier on me when dealing with multiple lists of issuez.
- store / update redirect to the feature page
rather than a standalone stories index.