Fix view Helper and writing to http headers multiple times
ExecuteTemplate, when writing to the http.ResponseWriter
sends the http header. I found this blog post online that
explains that we can write to a buffer to ensure the
template executes without an error, and based on the result
send our http header -
https://blog.questionable.services/article/approximating-html-template-inheritance/

- add send() function to the viewService struct that
writes the buffered template to the http.ResponseWriter
and sets the proper response code.