WIP - Add Admin + Roles, Capabilities and Permissions
This is a little roughed in. Only the admin endpoints
are guarded. The functionality is there to protect other
resource endpoints, though.
I drew inspiration from WordPress for capabilities.
I have experience with them, and the design seems to work
well for WP - a framework used heavily for membership
sites, etc.